bonus finish (need to get the tests)

+++ circle-01/ft_printf/srcs/process_nbr.c
@@ -18,9 +18,9 @@ t_bool	process_nbr(t_form *f, va_list ap)
 
 	ret = FL;
 	if (f->t == 4 || f->t == 5)
-		ret = prnt_dec(f, (long long)(va_arg(ap, int)), TR);
+		ret = prnt_dec(f, va_arg(ap, int), TR);
 	else if (f->t == 6)
-		ret = prnt_dec(f, (long long)(va_arg(ap, unsigned int)), FL);
+		ret = prnt_dec(f, va_arg(ap, unsigned int), FL);
 	else if (f->t == 7)
 		ret = prnt_hex(f, va_arg(ap, unsigned int), FL, FL);
 	else if (f->t == 8)
