+++ circle-01/ft_printf/srcs/process_nbr_h.c
@@ -6,7 +6,7 @@
 /*   By: jseo <jseo@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/18 18:17:11 by jseo              #+#    #+#             */
-/*   Updated: 2021/03/18 18:17:13 by jseo             ###   ########.fr       */
+/*   Updated: 2021/03/18 18:38:37 by jseo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,12 +18,12 @@ t_bool	process_nbr_h(t_form *f, va_list ap)
 
 	ret = FL;
 	if (f->t == 4 || f->t == 5)
-		ret = prnt_dec(f, (long long)(va_arg(ap, int)), TR);
+		ret = prnt_dec(f, (short)(va_arg(ap, int)), TR);
 	else if (f->t == 6)
-		ret = prnt_dec(f, (long long)(va_arg(ap, unsigned int)), FL);
+		ret = prnt_dec(f, (unsigned short)(va_arg(ap, int)), FL);
 	else if (f->t == 7)
-		ret = prnt_hex(f, va_arg(ap, unsigned int), FL, FL);
+		ret = prnt_hex(f, (unsigned short)(va_arg(ap, int)), FL, FL);
 	else if (f->t == 8)
-		ret = prnt_hex(f, va_arg(ap, unsigned int), TR, FL);
+		ret = prnt_hex(f, (unsigned short)(va_arg(ap, int)), TR, FL);
 	return (ret);
 }
