+++ circle-01/ft_printf/srcs/process_nbr_l.c
@@ -18,12 +18,12 @@ t_bool	process_nbr_l(t_form *f, va_list ap)
 
 	ret = FL;
 	if (f->t == 4 || f->t == 5)
-		ret = prnt_dec(f, (long long)(va_arg(ap, int)), TR);
+		ret = prnt_dec(f, va_arg(ap, long), TR);
 	else if (f->t == 6)
-		ret = prnt_dec(f, (long long)(va_arg(ap, unsigned int)), FL);
+		ret = prnt_dec(f, va_arg(ap, unsigned long), FL);
 	else if (f->t == 7)
-		ret = prnt_hex(f, va_arg(ap, unsigned int), FL, FL);
+		ret = prnt_hex(f, va_arg(ap, unsigned long), FL, FL);
 	else if (f->t == 8)
-		ret = prnt_hex(f, va_arg(ap, unsigned int), TR, FL);
+		ret = prnt_hex(f, va_arg(ap, unsigned long), TR, FL);
 	return (ret);
 }
